# backend/tracker/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Nutrition Tracker API
  version: 1.0.0
  description: |
    The Tracker API allows users to manage their products, consumption logs, nutritional statistics, shopping lists, and personal settings. 
    This API provides various endpoints for retrieving product details, adding or removing items from shopping lists, consuming products and tracking daily nutrition, 
    updating personal settings, and more.

    Key functionalities include:
    - Managing products: Get, create, and search products by ID or name.
    - Consuming products: Record product consumption and update nutrition statistics.
    - Managing shopping lists: Create, retrieve, and delete shopping lists, as well as add or remove items from them.
    - Tracking daily nutrition: Retrieve detailed nutrition statistics based on consumed products.
    - Personal settings: View and update user-specific nutritional goals and preferences.
    - Calendar integration: View, create, and update calendar events related to food consumption.
...
servers:
  - url: http://localhost:9401/tracker
    description: "standard url"
  - url: http://tracker-api:9401/tracker
    description: "url when fetching data from this api inside a docker network"
paths:
...
  /diary/date/{date}:
    get:
      summary: Retrieve the diary of consumed products for a specific date.
      description: Fetches the list of consumed products for a given date.
      tags:
        - diary
      operationId: getDiaryByDate
      parameters:
        - name: date
          in: path
          required: true
          description: The date for which to retrieve the diary of consumed products (format YYYY-MM-DD).
          schema:
            type: string
            format: date
      responses:
        200:
          description: Successfully retrieved the diary for the specified date.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyProductsConsumed'
        400:
          description: Invalid date format.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid date format. Use YYYY-MM-DD."
        404:
          description: No records found for the given date.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "No records found for this date."
        500:
          description: Internal server error when fetching diary data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to fetch diary entry."
  /diary/date/{date}/{id}:
    get:
      summary: Retrieve a specific consumed product by its ID.
      description: Fetches details of a consumed product based on its ID for a given date.
      tags:
        - diary
      operationId: getConsumedProductByID
      parameters:
        - name: date
          in: path
          required: true
          description: The date for which to retrieve the product (format YYYY-MM-DD).
          schema:
            type: string
            format: date
        - name: id
          in: path
          required: true
          description: The ID of the consumed product.
          schema:
            type: integer
      responses:
        200:
          description: Successfully retrieved the consumed product.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsumedProduct'
        400:
          description: Invalid ID format (must be an integer).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid ID format. Must be an integer."
        404:
          description: Product not found for the given ID.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Product not found."
        500:
          description: Internal server error while fetching the consumed product.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to fetch product."
    delete:
      summary: Delete a specific consumed product.
      description: Deletes a consumed product by its ID and updates nutrition statistics.
      tags:
        - diary
      operationId: deleteConsumedProduct
      parameters:
        - name: date
          in: path
          required: true
          description: The date for which the product is being deleted (format YYYY-MM-DD).
          schema:
            type: string
            format: date
        - name: id
          in: path
          required: true
          description: The ID of the consumed product to delete.
          schema:
            type: integer
      responses:
        200:
          description: Successfully deleted the product and updated nutrition statistics.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Product deleted successfully."
        400:
          description: Invalid ID format (must be an integer).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid ID format. Must be an integer."
        404:
          description: Product not found for the given ID.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Product not found."
        500:
          description: Internal server error when deleting the product or updating nutrition statistics.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to delete product."
...
  /calendar/{date}:
    get:
      summary: Retrieve calendar entries for a specific date.
      description: Fetches calendar entries for a given date.
      tags:
        - calendar
      operationId: getCalendarByDate
      parameters:
        - name: date
          in: path
          required: true
          description: The date for which to retrieve calendar entries (format YYYY-MM-DD).
          schema:
            type: string
            format: date
      responses:
        200:
          description: Successfully retrieved calendar entries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CalendarDTO'
        400:
          description: Invalid date format.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid date format. Use YYYY-MM-DD."
        500:
          description: Internal server error while fetching calendar entries.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to fetch calendar entries."
  /calendar:
    post:
      summary: Add a new calendar entry.
      description: Creates a new calendar entry with date and time details.
      tags:
        - calendar
      operationId: addCalendarEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarDTO'
      responses:
        201:
          description: Successfully added calendar entry.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Calendar entry added successfully."
        400:
          description: Invalid input data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid request payload."
        500:
          description: Internal server error while adding calendar entry.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to add calendar entry."
  /calendar/{id}:
    patch:
      summary: Update an existing calendar entry.
      description: Modifies details of a calendar entry by ID.
      tags:
        - calendar
      operationId: updateCalendarEntry
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the calendar entry to update.
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarDTO'
      responses:
        200:
          description: Successfully updated the calendar entry.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Calendar entry updated successfully."
        400:
          description: Invalid input data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid request payload."
        500:
          description: Internal server error while updating calendar entry.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to update calendar entry."

    delete:
      summary: Delete a calendar entry.
      description: Removes a calendar entry by ID.
      tags:
        - calendar
      operationId: deleteCalendarEntry
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the calendar entry to delete.
          schema:
            type: integer
      responses:
        200:
          description: Successfully deleted the calendar entry.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Calendar entry deleted successfully."
        404:
          description: Calendar entry not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Calendar entry not found."
        500:
          description: Internal server error while deleting calendar entry.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to delete calendar entry."
...
components:
  schemas:
...
    DailyProductsConsumed:
      type: object
      properties:
        ID:
          type: integer
        Date:
          type: string
        Products:
          type: array
          items:
            $ref: '#/components/schemas/ConsumedProduct'
    ConsumedProduct:
      type: object
      properties:
        ID:
          type: integer
          format: uint
        DailyProductsConsumedID:
          type: integer
        ProductID:
          type: integer
        ProductName:
          type: string
        Brand:
          type: string
        Category:
          type: string
          enum: [Frühstück, Mittagessen, Abendessen, Snack]
        WeightInGrams:
          type: integer
        Calories:
          type: integer
        ProteinsInGrams:
          type: number
          format: float
        FatsInGrams:
          type: number
          format: float
        CarbsInGrams:
          type: number
          format: float
...
    CalendarDTO:
      type: object
      properties:
        ID:
          type: integer
        Date:
          type: string
        BeginTime:
          type: string
        EndTime:
          type: string
        Name:
          type: string
        Description:
          type: string
...